requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
Django==2.2
django-countries==6.1.3
django-redis==4.12.1
django-stdimage==5.1.1
djangorestframework==3.11.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
Pillow==7.2.0
progressbar2==3.52.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-utils==2.4.0
pytz==2020.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.1.0
